+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/engine_functions/logs_components/fluentbit/factory.go
@@ -1,7 +1,10 @@
 package fluentbit
 
-func CreateFluentbitConfiguredForKurtosis() *Fluentbit {
-	config := newDefaultConfigForKurtosisCentralizedLogs()
+func CreateFluentbitConfiguredForKurtosis(
+	logsDatabaseHost string,
+	logsDatabasePort uint16,
+) *Fluentbit {
+	config := newDefaultConfigForKurtosisCentralizedLogsForDocker(logsDatabaseHost, logsDatabasePort)
 	fluentbit := NewFluentbit(config)
 	return fluentbit
 }