+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/engine_functions/logs_components/logs_collector.go
@@ -7,9 +7,13 @@ import (
 
 type LogsCollector interface {
 	GetPrivateTcpPortSpec() (*port_spec.PortSpec, error)
+	GetPrivateHttpPortSpec() (*port_spec.PortSpec, error)
 	GetContainerArgs(
 		containerName string,
 		containerLabels map[string]string,
+		volumeName string,
 		networkId string,
+		dockerManager *docker_manager.DockerManager,
 	) (*docker_manager.CreateAndStartContainerArgs, error)
+	WaitForAvailability() error
 }