+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/engine_functions/logs_components/logs_database.go
@@ -13,5 +13,4 @@ type LogsDatabase interface {
 		volumeName string,
 		networkId string,
 	) (*docker_manager.CreateAndStartContainerArgs, error)
-	GetConfigContent() (string, error)
 }